Use Endpoint CRD to configure routes (#270)
* Add Global CIDR to Cluster CRD

There is a growing need to be able to support multi cluster networking
across brownfield deployments that have limited abilities to create non
overlapping CIDRs for their Pod(s) and Service(s) networks. As discussed
during some meetings this can be addressed by having a Global multi
cluster network that is non-overlapping and can do NATting while still
be able to leverage the advantages that Submariner brings.

This patch adds the ability to define a Global CIDR segment associated
with a specific cluster and store the information as part of the
cluster's Cluster CRD. The autogenerated part was added by executing
`./scripts/codegen`

* Use Endpoint CRD to configure routes

Currently the `submariner-agent` makes use of `Cluster` CRDs to set
up the routes. This, however, is information that is more natural to
the `Endpoint` CRD as it describes the subnets that are to be found
when connecting that endpoint. Moreover, having a single source of
truth instead of two simplifies the code and allows for easier
integration with `submariner-globalnet`.

This patch makes the `Endpoint` be the source of truth for setting up
routes. By doing so, there is no need for the `submariner-routeagent` to
monitor `Cluster`s anymore and therefore the code that is related has
been removed, simplifying the codebase in consequence.