Report startup errors in the Gateway StatusFailure
Aded a SetGatewayStatusError method to the Gateway syncer that creates/updates
the Gateway resource with a specific StatusFailure. This is called prior to
exiting via Fatalf.

This doesn't include some bootstrap failures that prevent the syncer from being
created, like failure to create the kube client.

Eliminated a couple error paths that never actually result in an error to simplify:

 - util.GetLocalCluster - this function siumply copies fields so I eliminated it
   altogether and inlined it.

 - cableengine.NewEngine - removed the error return - any errors should be reported by
   StartEngine..

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>