Ensure Gateway resource is deleted on shutdown
The submariner Gateway resource is deleted by the syncer when
the syncGatewayStatus loop exits but it's run asynchronously
from the main function so, due to timing, it may not get deleted
on shutdown. This can result in an intermittent gateway failover
E2E failure b/c, after deleting the active gateway pod, it ensures
the Gateway is fully connected, assuming it's a new instance. But,
if the prior Gateway was not deleted, then it will observe the fully
connected state of the prior stale Gateway. This can lead to
subsequent timing failures where it is assumed the gateway is fully
active.

Ensure that the Gateway is deleted by runnng the Gateway syncer
synchronously from main and waiting for it to complete on shutdown.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>