Verify new peer config in WG driver ConnectToEndpoint
This addresses a TODO: "verify configuration". It's not clear exactly
what the original author intended but it seems reasonable it was to
verify the relevant data in the PeerConfig matches the equivalent
data in the newly created Peer. Previously it only logged an error
if the new Peer wasn't found so the same was done if the data
doesn't match.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>